Description: changed data model type int to int64 Reviewed By: Mandar Change Log: Bug ID:

+++ trunk/technical/src/appserver/src/opensoach.com/models/cache.go
@@ -2,9 +2,9 @@ package models
 
 type UserSessionInfo struct {
 	UserID     int64    `json:"userid"`
-	CpmID      int      `json:"cpmid"`
-	UserRoleID int      `json:"userroleid"`
-	CustomerID int      `json:customerid`
+	CpmID      int64    `json:"cpmid"`
+	UserRoleID int64    `json:"userroleid"`
+	CustomerID int64    `json:customerid`
 	UserType   int      `json:"usertype"`
 	ModDB      ConfigDB `json:"moddb"`
 }
